Cloak graphics trailing from the hands.
I've achieved a nice effect by following the cloak down from the
hands in continuous strands of material, with a few differently
coloured strands woven in to give the impression of folds.

This strand process is random but I wanted to keep cloak graphics
consistent between frames (unless the user moved). To do this I've
implemented a standalone PRNG and keep track of the seed for the
cloak.

If I don't want the cloak to change I can just replay with the same
seed.